Add tests for sub,silent mutation, and invalid queries
